Create What do I Wear?
Use the sizing chart provided, create a series of logical expressions that prints the size of a t-shirt based on the measurements of shirtWidth, shirtLength, and shirtSleeve. Valid sizes include S, M, L, XL, 2XL, and 3XL.
Then print N/A to the console because the measurements don't all match up with one particular size.